--- sources/packages/web/src/types/contracts/EducationContract.ts
@@ -1,4 +1,0 @@
-export interface EducationProgram {
-  name: string;
-  code?: string;
-}
