+++ sources/packages/web/src/types/contracts/EducationProgram.ts
