+++ sources/packages/web/src/types/index.ts
@@ -2,13 +2,12 @@ export * from "@/types/AppRoutes";
 export * from "@/types/AuthStatus";
 export * from "@/types/ApplicationToken";
 export * from "@/types/contracts/ConfigContract";
-export * from "@/types/contracts/EducationContract";
 export * from "@/types/contracts/StudentContract";
 export * from "@/types/contracts/ProgramYearContract";
 export * from "@/types/contracts/UserContract";
 export * from "@/types/contracts/institution/InstitutionLocation";
 export * from "@/types/contracts/institution/InstitutionUser";
-export * from "@/types/contracts/institution/EducationProgram";
+export * from "@/types/contracts/EducationProgram";
 export * from "@/types/contracts/institution/EducationProgramOfferingDto";
 export * from "@/types/contracts/OfferingContact";
 export * from "@/types/StoreState";
